docs: make announcement examples more discoverable
