+++ packages/react/src/Banner/Banner.examples.stories.tsx
@@ -2,10 +2,7 @@ import {Banner} from '../Banner'
 import {action} from '@storybook/addon-actions'
 import Link from '../Link'
 import type {Meta} from '@storybook/react'
-import {AriaAlert, AriaStatus} from '../live-region'
-import FormControl from '../FormControl'
-import RadioGroup from '../RadioGroup'
-import Radio from '../Radio'
+import {AriaAlert} from '../live-region'
 import {Button} from '../Button'
 import React from 'react'
 import {useFocus} from '../internal/hooks/useFocus'
@@ -42,49 +39,6 @@ export const WithUserAction = () => {
       >
         Update profile
       </Button>
-    </>
-  )
-}
-
-export const WithDynamicContent = () => {
-  type Choice = 'one' | 'two' | 'three'
-  const messages: Map<Choice, string> = new Map([
-    ['one', 'This is a message for choice one'],
-    ['two', 'This is a message for choice two'],
-    ['three', 'This is a message for choice three'],
-  ])
-  const [selected, setSelected] = React.useState<Choice>('one')
-
-  return (
-    <>
-      <Banner
-        title="Info"
-        description={<AriaStatus>{messages.get(selected)}</AriaStatus>}
-        onDismiss={action('onDismiss')}
-        primaryAction={<Banner.PrimaryAction>Button</Banner.PrimaryAction>}
-        secondaryAction={<Banner.SecondaryAction>Button</Banner.SecondaryAction>}
-      />
-      <RadioGroup
-        sx={{marginTop: 4}}
-        name="options"
-        onChange={selected => {
-          setSelected(selected as Choice)
-        }}
-      >
-        <RadioGroup.Label>Choices</RadioGroup.Label>
-        <FormControl>
-          <Radio value="one" defaultChecked />
-          <FormControl.Label>Choice one</FormControl.Label>
-        </FormControl>
-        <FormControl>
-          <Radio value="two" />
-          <FormControl.Label>Choice two</FormControl.Label>
-        </FormControl>
-        <FormControl>
-          <Radio value="three" />
-          <FormControl.Label>Choice three</FormControl.Label>
-        </FormControl>
-      </RadioGroup>
     </>
   )
 }
