+++ packages/react/src/Banner/Banner.features.stories.tsx
@@ -4,6 +4,10 @@ import {action} from '@storybook/addon-actions'
 import type {Meta} from '@storybook/react'
 import {Banner} from '../Banner'
 import Link from '../Link'
+import FormControl from '../FormControl'
+import RadioGroup from '../RadioGroup'
+import Radio from '../Radio'
+import {AriaStatus} from '../live-region'
 
 const meta = {
   title: 'Experimental/Components/Banner/Features',
@@ -256,5 +260,48 @@ export const CustomIcon = () => {
       onDismiss={action('onDismiss')}
       variant="upsell"
     />
+  )
+}
+
+export const WithAnnouncement = () => {
+  type Choice = 'one' | 'two' | 'three'
+  const messages: Map<Choice, string> = new Map([
+    ['one', 'This is a message for choice one'],
+    ['two', 'This is a message for choice two'],
+    ['three', 'This is a message for choice three'],
+  ])
+  const [selected, setSelected] = React.useState<Choice>('one')
+
+  return (
+    <>
+      <Banner
+        title="Info"
+        description={<AriaStatus>{messages.get(selected)}</AriaStatus>}
+        onDismiss={action('onDismiss')}
+        primaryAction={<Banner.PrimaryAction>Button</Banner.PrimaryAction>}
+        secondaryAction={<Banner.SecondaryAction>Button</Banner.SecondaryAction>}
+      />
+      <RadioGroup
+        sx={{marginTop: 4}}
+        name="options"
+        onChange={selected => {
+          setSelected(selected as Choice)
+        }}
+      >
+        <RadioGroup.Label>Choices</RadioGroup.Label>
+        <FormControl>
+          <Radio value="one" defaultChecked />
+          <FormControl.Label>Choice one</FormControl.Label>
+        </FormControl>
+        <FormControl>
+          <Radio value="two" />
+          <FormControl.Label>Choice two</FormControl.Label>
+        </FormControl>
+        <FormControl>
+          <Radio value="three" />
+          <FormControl.Label>Choice three</FormControl.Label>
+        </FormControl>
+      </RadioGroup>
+    </>
   )
 }
